replace input connection to package
